Fix invalid plugins not being hidden in some cases
When an invalid plugin loaded after all valid plugins in the initial
load order, its absence would not be noticed and it would not have its
card removed. Also, if the only change was to remove an invalid plugin,
that would still be reported as a change, even if it could not be
removed in practice (e.g. in Oblivion), so re-sorting would show the
same 'change' over and over.

Instead, LOOT no longer counts removing a plugin as changing the load
order, and always removes any corresponding plugin cards.